requirements.txt
streamlit>=1.24.0
python-dotenv>=1.0.0
requests>=2.31.0
PyYAML>=6.0.1
pandas>=2.0.3
aiohttp>=3.8.5
python-dateutil>=2.8.2
typing-extensions>=4.7.1
colorama>=0.4.6
watchdog>=3.0.0 